requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
firebase-admin
langchain-google-genai
langchain
